Add support to unsigned long fields
This commit enhances the Getter.hpp file to support the
retrieval of unsigned long fields with getField(), which
was previously not supported.
This can be useful for ETSI ITS-G5 messages, which are
relying on ASN.1 and contain a StationID, which is translated
by asn1c into an unsigned long type.